attempted to link the user from auth0 with the chat, but to no prevail, may need to be a stretch goal
